requirements.txt
numpy
pandas
matplotlib
seaborn
torch
scikit-learn
joblib
tensorflow
pyserial
ipykernel
